회원 탈퇴시, 수정시, 회원 정보보기 시 권한 검증

+++ src/main/java/com/example/prj1be/controller/MemberController.java
@@ -5,7 +5,6 @@ import com.example.prj1be.service.MemberService;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.context.request.WebRequest;
@@ -66,16 +65,33 @@ public class MemberController {
 
 
     @GetMapping
-    public ResponseEntity<Member> view(String id) {
+    public ResponseEntity<Member> view(String id,
+                                       @SessionAttribute(value = "login", required = false) Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if(!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         Member member = service.getMember(id);
 
         return ResponseEntity.ok(member);
     }
 
     @DeleteMapping
-    public ResponseEntity delete(String id) {
-        //TODO : 로그인 했는지? -> 안했으면 401
-        //TODO : 자기 정보인지? -> 아니면 403
+    public ResponseEntity delete(String id,
+                                 @SessionAttribute(value = "login", required = false) Member login ) {
+
+        if( login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); //401
+        }
+
+        if( !service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); //403
+        }
+
 
         if (service.deleteMember(id)) {
             return ResponseEntity.ok().build();
@@ -84,8 +100,16 @@ public class MemberController {
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Member member) {
-        // TODO: 로그인 했는지? 자기 정보인지?
+    public ResponseEntity edit(@RequestBody Member member,
+                               @SessionAttribute(value = "login", required = false) Member login) {
+
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); //401
+        }
+
+        if(!service.hasAccess(member.getId(), login))  {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
 
         if (service.update(member)) {
             return ResponseEntity.ok().build();
